updated_requirements.txt
flask
langchain-community
langchain-experimental
transformers
faiss-cpu
datasets